requirements.txt
# Runtime dependencies
earthengine-api>=1.0.0
click>=8.1.0
pyyaml>=6.0
geopy>=2.4.0
rasterio>=1.3.0
requests>=2.31.0
diskcache>=5.6.0
numpy>=1.24.0
scikit-image>=0.21.0
scipy>=1.10.0
opencv-python-headless>=4.8.0
matplotlib>=3.8.0
jinja2>=3.1.0
pillow>=10.0.0
rich>=13.0.0
